+++ TestApp/InputParser.cs
@@ -64,8 +64,8 @@ namespace TestApp
             if (rngPos != -1)
             {
                 if (
-                    uint.TryParse(input[..rngPos], out var lhs) &&
-                    uint.TryParse(input[(rngPos + 1)..stepPos], out var rhs) &&
+                    ushort.TryParse(input[..rngPos], out var lhs) &&
+                    ushort.TryParse(input[(rngPos + 1)..stepPos], out var rhs) &&
                     lhs <= rhs
                 )
                 {
@@ -82,7 +82,7 @@ namespace TestApp
             }
             else
             {
-                if (uint.TryParse(input[..stepPos], out var val))
+                if (ushort.TryParse(input[..stepPos], out var val))
                 {
                     result = new SingularInput(val);
                 }
@@ -97,7 +97,7 @@ namespace TestApp
                 return result;
             }
 
-            if (stepPos is var _ && (uint.TryParse(input[(stepPos + 1)..], out var stepBy) && stepBy > 0))
+            if (stepPos is var _ && ushort.TryParse(input[(stepPos + 1)..], out var stepBy) && stepBy > 0)
             {
                 return new StepByInput(result, stepBy);
             }
